Fix private shadow memory by recording the private base (#18253)
Before this patch, to map private buffer into shadow memory, we use the
lower bits of private address as offset
```
shadow_ptr = private_shadow_base + (addr & (MSAN_PRIVATE_SIZE - 1));
```
This is based on assumption that the high bits of allocated private
address for each workitem is same.
But recently, we observed that this is not true.

Therefore, we record the base address of private memory for each
workitem, so that we can get the offset.
```
shadow_ptr = private_shadow_base + (addr - private_base[workitem_id]);
```

Misc.
- Fix warning in AddressSanitizer
- Change the return type of "__msan_get_shadow" to remove "ptrtoint"
instruction