Use abort instead of exit when in error (#19085)
When we try to exit the problem when reporting the error, the `exit(1)`
could cause hangs in SYCL runtime because it skips some processes before
the SYCL shutdown process in atexit stage. The `abort()` is more
accurate here because it would stop the whole process right away, and no
more SYCL shutdown process with the unstable, early exited program.