Fix creation of context with parent device and its sub-devices (#19223)
This use case seems to be allowed. For example, according to:
https://intel.github.io/llvm/MultiTileCardWithLevelZero.html#context
"Both root-devices and sub-devices can be within single context, but
they all should be of the same SYCL platform."

Before change CollectDevicesAndSubDevices was resulting in duplicate
devices being returned which leads to an error in
https://github.com/intel/llvm/blob/6af08fe9c6fd4dc8e433555646606898c71d92fb/unified-runtime/source/common/ur_pool_manager.hpp#L178-L179
because of the duplicate pool descriptors.