Add DX11 memory interop (#19217)
In DX11 Texture1D and Texture3D cannot be shared between process or
devices, so only Texture2D is used but its layout is adapted (height =
1 for 1D and ArraySlices = depth for 3D) for the interop purposes and
testing 1D and 3D image operations on it in the SYCL kernel.
The new DXGI adapter selection fixes issues with non-matching devices
between DX and SYCL which also failed the DX12 tests when more than 1
potential adapter is visible to DirectX. Ideally the introduction of
LUIDs to SYCL will resolve that completely when they are properly
matched. The future flow should change to - 1) create SYCL device (so
the device selection itself can be manipulated via
`ONEAPI_DEVICE_SELECTOR`) , 2) get DXGI adapter for creating a DX
logical device (D3D11 or D3D12) by matching adapter LUIDs (this is going
to depend on a LUID device info query extension for SYCL).

Some notes on synchronisation:

- IKeyedMutex is required for synchronising the access to the shared
resource (texture) between devices or processes.
- Currently the SYCL queue calls wait after submission to execute
immediately and block until completion, but we can use ID3D11Fence
imported in SYCL to signal the completion of the work to D3D11 via SYCL
in the future when this kind of interop is considered.