Don't check nvmlShutdown return in destructor (#19512)
`UR_CHECK_ERROR` throws an exception so we can't use it from a
destructor.

Since devices only get released at application shutdown, there's not
much we can do if `nvmlShutdown` reports an error, and the OS will
promptly cleanup anything left as it tears down the process anyway.

So just skip checking the error code from `nvmlShutdown`, this is
already what we do for other cleanup/shutdown in this destructor.

Fixes: https://github.com/intel/llvm/issues/19500