Simplify enqueue. (#19550)
We were creating excessive numbers of threads. When we know we want a
given amount of threads, just divide the number of workgroups by the
number of threads and have each thread process that many workgroups.

This implementation also means we no longer need to resize workgroups,
which was not generally safe.