Support preinstalled GTest (#19720)
Don't fetch and build it if a compatible version is installed on the
system.

GMock is considered part of GTest to CMake (and not even a component)
but in some Linux distros its a separate package, so we need to handle
the case where one is installed but not the other.

Issue: https://github.com/intel/llvm/issues/19635

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>