Fix OpenCL target error on rebuild (#19740)
The problem was `OpenCL_FOUND` was getting set to `TRUE` on the second
invocation of the module because `find_package` was finding the
`FetchContent` clone in `SOURCE_TREE/build/__deps`.

We use `OpenCL_FOUND` to specifically refer to external system-wide
installs, so use `NO_CMAKE_PACKAGE_REGISTRY` to make sure we don't find
FetchContent clones.

Also not related to the fix here but re-order `FetchContent_Declare`
`FetchContent_GetProperties`, the `Declare` should come first and we
`MakeAvailable` should not be inside the `if`, we only check if it's
populated to print the message.

Closes: https://github.com/intel/llvm/issues/19737

Signed-off-by: Sarnie, Nick <nick.sarnie@intel.com>