Fix platform::get_devices (#19810)
In relation to the clarification in
https://github.com/KhronosGroup/SYCL-Docs/pull/861, the implementation
of `platform::get_devices` does not work correctly for `custom` and
`automatic`. This commit fixes this by ensuring these return the devices
with custom type and the default device respectively.

Additionally, it adds the `UR_DEVICE_TYPE_CUSTOM` to UR and adds proper
handling for `UR_DEVICE_TYPE_DEFAULT` in the UR OpenCL adapter.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>