Remove fallback assertions (#18310)
This commit removes the fallback assertion implementation for SYCL
kernels. The expected behavior after this is that backends that do not
support native asserts, as reported through the ext_oneapi_native_assert
aspect, will ignore assertions in kernel code.

This commit also makes the L0 adapter report support for native assert,
as it was incorrectly reporting false before.

---------

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>
Co-authored-by: Alexey Bader <alexey.bader@intel.com>