Fix and enable bindless image copy tests on L0 (#20096)
The `sycl_ext_oneapi_bindless_images` extension specification documents
multiple overloads for copying between bindless images depending on
source and destination types (i.e. image mem handle to USM, USM to USM,
etc.).

At Unified Runtime level, there is a single function
`urBindlessImagesImageCopyExp` which implements all kinds of copies.
However, at lower levels such as CUDA and Level Zero there are several
different APIs for different kinds of copies. To understand which one to
call, a "copy direction" argument is passed.

However, that "copy direction" information is useless for the L0 path,
because L0 API is not based on a direction, but it is based on input
types (mem handle vs USM). What we have today is tailored to CUDA and if
we change copy directions to be tailored to L0, then CUDA breaks
(because input types also matter there, but to lesser extent).

In order to fix all issues with bindless image copies and make them work
across all adapters, a new argument is introduced into
`urBindlessImagesImageCopyExp` to specify types of source and
destination arguments. L0 UR adapter makes a decision about which API to
call solely on this information, ignoring "copy direction".

CUDA & HIP adapters are left as-is - they still use "copy direction"
information and sometimes try to guess the type of input arguments to
call the right API. With the new argument introduced by this patch,
those adapters could be refactored to reduce amount of guessing, but
that is outside of the scope of this PR.