Set pointer kernel arguments only for queue's associated device (#20177)
In multi-device context, ensure that pointer kernel arguments are set
only for the device associated with the queue being used for kernel
launch. Previously, arguments were set for all devices in the kernel's
device map, which was unnecessary and potentially incorrect when
launching on a specific device.

Same will be done for v2 adapter here:
https://github.com/intel/llvm/pull/20179